Reorganise router module
